Move OnChangeCommand to commands module
* Along with DotWritingProgressMonitor, OnCommand
* Fix build.sbt to no build "onchange" any more
