+++ commands/src/main/resources/META-INF/services/com.reactific.riddl.commands.CommandPlugin
@@ -4,3 +4,4 @@ com.reactific.riddl.commands.ParseCommand
 com.reactific.riddl.commands.RepeatCommand
 com.reactific.riddl.commands.StatsCommand
 com.reactific.riddl.commands.ValidateCommand
+com.reactific.riddl.commands.OnChangeCommand
