+++ commands/src/main/scala/com/reactific/riddl/commands/DotWritingProgressMonitor.scala
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.reactific.riddl.translator.onchange
+package com.reactific.riddl.commands
 
 import com.reactific.riddl.language.CommonOptions
 import com.reactific.riddl.utils.Logger
