--- onchange/src/main/resources/META-INFo/services/com.reactific.riddl.commands.CommandPlugin
@@ -1,1 +1,0 @@
-com.reactific.riddl.hugo_git_check.GitCheckCommand
